backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
alembic
yfinance
pandas
pandas-ta
scikit-learn
xgboost
python-dotenv
joblib
